+++ lib/generators/rspec/scaffold/templates/index_spec.rb
@@ -22,7 +22,7 @@ describe "<%= table_name %>/index.html.<%= options[:template_engine] %>" do
 <% if webrat? -%>
     rendered.should have_selector("tr>td", :content => <%= value_for(attribute) %>.to_s, :count => 2)
 <% else -%>
-    # Run the generator again with the --webrat-matchers flag if you want to use webrat matchers
+    # Run the generator again with the --webrat flag if you want to use webrat matchers
     assert_select "tr>td", :text => <%= value_for(attribute) %>.to_s, :count => 2
 <% end -%>
 <% end -%>