+++ lib/rspec/rails/example/helper_example_group.rb
@@ -61,6 +61,7 @@ module RSpec::Rails
 
     included do
       metadata[:type] = :helper
+
       before do
         controller.controller_path = _controller_path
       end